backend/requirements.txt
openai
pinecone-client
pdfminer.six
Flask
python-dotenv